requirements.txt
Shapely==1.5.13
Flask==0.10.1
matplotlib==1.5.1
scipy==0.19.0
plumbum==1.6.2
numpy==1.12.1
ipython==6.1.0
Pillow==4.2.1
keras_resnet